move formatDate to DatePicker folder
